Removed awaits. Improved AFSClient API.
Removed awaits because some consumers are using older compilers that cannot support the await keyword.
These consumers were trying to synchronously wait for tasks to finish which was causing deadlocks in UI apps.
Changed AFSClient ctor that takes connection string to be a static factory method, because it was conflicting with a more
useful ctor overload.

Made Urls public and exposed the DefaultBaseAddress.